move xss filter
